Relocate_and_add_actions
